スポット追加フォーム切り分け
