+++ frontend/src/components/SpotList.tsx
@@ -1,21 +1,9 @@
 import React, { useState } from 'react'
-import { Link, useNavigate } from 'react-router-dom'
-import axios from 'axios'
-import {
-  Button,
-  Form,
-  Input,
-  TimePicker,
-  Select,
-  Space,
-  InputNumber,
-} from 'antd'
-const { Option } = Select
+import { Button } from 'antd'
 import { PlusOutlined } from '@ant-design/icons'
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react'
-
-const apiUrl = process.env.REACT_APP_API_URL
+import SpotCreate from './SpotCreate'
 
 type Plan = {
   id: number
@@ -32,22 +20,12 @@ type Props = {
 }
 
 const SpotList: React.FC<Props> = ({ plan, getTrip }) => {
-  const [data, setData] = useState<any>({ plan_id: plan.id })
-
-  async function addSpot() {
-    try {
-      const res = await axios.post(`${apiUrl}/spots`, data)
-      console.log(res)
-      getTrip()
-    } catch (error) {
-      console.log(error)
-    }
-  }
+  const [showCreate, setShowCreate] = useState<boolean>(false)
 
   const styles = {
     spots: css`
       width: 100%;
-      margin-bottom: 20px;
+      margin-bottom: 10px;
     `,
     spot: css`
       padding: 5px 10px;
@@ -81,7 +59,7 @@ const SpotList: React.FC<Props> = ({ plan, getTrip }) => {
     form: css`
       margin-bottom: 20px;
       text-align: center;
-    `
+    `,
   }
 
   return (
@@ -106,67 +84,21 @@ const SpotList: React.FC<Props> = ({ plan, getTrip }) => {
             )
           })}
       </div>
-      <Form onFinish={addSpot} css={styles.form}>
-        <div style={{ display: 'flex' }}>
-          <TimePicker
-            placeholder='開始時間'
-            format='HH:mm'
-            size='small'
-            onChange={(time, timeString) =>
-              setData({ ...data, start_time: timeString })
-            }
-          />
-          <TimePicker
-            placeholder='終了時間'
-            format='HH:mm'
-            size='small'
-            onChange={(time, timeString) =>
-              setData({ ...data, end_time: timeString })
-            }
-          />
-          <Select
-            placeholder='カテゴリーを選択'
-            size='small'
-            onChange={(event) => setData({ ...data, category_id: event })}
-          >
-            <Option value='1'>新幹線</Option>
-            <Option value='2'>飛行機</Option>
-            <Option value='3'>車</Option>
-          </Select>
-        </div>
 
-        <div style={{ display: 'flex' }}>
-          <Input
-            placeholder='スポット名'
+      {!showCreate && (
+        <div style={{ textAlign: 'center', marginBottom: '20px' }}>
+          <Button
+            shape='circle'
+            icon={<PlusOutlined />}
             size='small'
-            onChange={(event) => setData({ ...data, name: event.target.value })}
-          />
-          <InputNumber
-            placeholder='金額'
-            size='small'
-            onChange={(event) => setData({ ...data, fee: event })}
+            onClick={() => setShowCreate(true)}
           />
         </div>
+      )}
 
-        <div style={{ display: 'flex' }}>
-          <Input
-            placeholder='リンク'
-            size='small'
-            onChange={(event) => setData({ ...data, link: event.target.value })}
-          />
-          <Input
-            placeholder='メモ'
-            size='small'
-            onChange={(event) => setData({ ...data, memo: event.target.value })}
-          />
-        </div>
-        <Button
-          shape='circle'
-          icon={<PlusOutlined />}
-          size='small'
-          htmlType='submit'
-        />
-      </Form>
+      {showCreate && (
+        <SpotCreate plan={plan} getTrip={getTrip} setFlag={setShowCreate} />
+      )}
     </>
   )
 }
